Moving classes into appropriate packages
