+++ src/main/java/cz/sparko/boxitory/controller/DownloadController.java
@@ -1,6 +1,6 @@
 package cz.sparko.boxitory.controller;
 
-import cz.sparko.boxitory.conf.NotFoundException;
+import cz.sparko.boxitory.exception.NotFoundException;
 import cz.sparko.boxitory.model.BoxStream;
 import cz.sparko.boxitory.service.BoxRepository;
 import org.slf4j.Logger;
